renamed visualizer components per new naming standard
