--- src/summary/NarrativeComponent.css
@@ -1,7 +1,0 @@
-span.missing-data {
-    border-bottom: 1px solid red;
-}
-
-span.structured-data {
-    border-bottom: 1px solid blue;
-}
