+++ src/summary/NarrativeNameValuePairsVisualizer.css
@@ -1,0 +1,7 @@
+span.missing-data {
+    border-bottom: 1px solid red;
+}
+
+span.structured-data {
+    border-bottom: 1px solid blue;
+}
