+++ src/summary/TabularNameValuePairsVisualizer.css
